requirements.txt
ultralytics
opencv-python
numpy
pandas
pillow
torch
faiss-cpu
fastapi
uvicorn
pyspark
